Firebase Auth working properly and nav bars updated
I have the sign in, sign out, and sign up functionality working properly now. I have setup the dynamic nav bars in the home and about pages as well and updated the nav bars on the other pages. Remaining piece is to have the app send the user to home page after signing in or signing out.